+++ app/partners/aipr/aipr-content.jsx
@@ -1,0 +1,12 @@
+"use client";
+
+import Layout from '@/components/Layout';
+import AIPR from '@/views/AIPR';
+
+export default function AIPRContent() {
+  return (
+    <Layout headerVariant="solid" bannerPreset={false}>
+      <AIPR />
+    </Layout>
+  );
+}